[transceiver/eeprom]: Move EEPROM helpers out of the test module
Extract the EEPROM parsing/comparison machinery and the recovery
verifiers into eeprom_content.py, a plain helper module matching the
existing datapath.py convention, so they are no longer imported from a
test module. test_eeprom_content.py keeps only its two test cases.

Helpers crossing the module boundary are now public
(run_per_port_eeprom_check, run_bulk_eeprom_check, parse_via_sfputil).

Addresses review feedback on PR #26915.
